Update annotation.rb
According to http://iiif.io/api/presentation/2.0/#image-resources, "the URI of the canvas must be repeated in the on field of the Annotation". This patch adds the "on" field to the annotation class and makes it a required field. 

It's worth noting that without this patch, https://github.com/IIIF/presentation-validator will fail the validation for an O'Sullivan generated manifest that contains an annotation. With the patch, it will succeed.